refactor: DAO 생성
